make tests run (but failing atm)
