+++ app/package.json
@@ -11,8 +11,8 @@
     "prepublish": "cd .. && pnpm turbo run build --scope=@previewjs/app --no-deps --include-dependencies",
     "build": "rimraf dist && pnpm tsc && pnpm tsc -p client/tsconfig.json && vite build client",
     "client:dev": "vite client",
-    "e2e-test": "ts-node run-test-suite-parallel.ts",
-    "e2e-test:group": "ts-node run-test-suite.ts",
+    "e2e-test": "ts-node --transpileOnly run-test-suite-parallel.ts",
+    "e2e-test:group": "ts-node --transpileOnly run-test-suite.ts",
     "lint": "eslint --ext .js,.jsx,.ts,.tsx ."
   },
   "eslintConfig": {
