+++ app/testing/helpers/app-controller.ts
@@ -2,6 +2,7 @@ import type { Preview, Workspace } from "@previewjs/core";
 import fs from "fs-extra";
 import path from "path";
 import playwright from "playwright";
+import "../../window";
 
 export class AppController {
   private preview: Preview | null = null;
