--- pro/src/tests/samples/.gitignore
@@ -1,3 +1,0 @@
-tmp/
-__previewjs__
-preview.config.js
