--- pro/src/tests/samples/default/package.json
@@ -1,11 +1,0 @@
-{
-  "dependencies": {
-    "@previewjs/config": "^0.1.0",
-    "@types/react": "^17.0.27",
-    "@types/react-dom": "^17.0.9",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-scripts": "4.0.3",
-    "typescript": "^4.4.3"
-  }
-}
