--- pro/src/tests/samples/default/src/App.tsx
@@ -1,40 +1,0 @@
-import React from "react";
-import "./App.css";
-import { Dependency } from "./Dependency";
-import logo from "./logo.svg";
-import { Foo } from "./types";
-
-interface AppProps {
-  children: React.ReactNode;
-  foo: Foo;
-  complex: React.DetailedHTMLProps<
-    React.AnchorHTMLAttributes<HTMLAnchorElement>,
-    HTMLAnchorElement
-  > & {
-    bar: "hi!";
-  };
-}
-
-function App(props: AppProps) {
-  return (
-    <div className="App">
-      <header className="App-header">
-        <img src={logo} className="App-logo" alt="logo" />
-        <p>
-          Edit <code>src/App.tsx</code> and save to reload.
-        </p>
-        <a
-          className="App-link"
-          href="https://reactjs.org"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          Learn React
-        </a>
-        <Dependency />
-      </header>
-    </div>
-  );
-}
-
-export { App };
