--- pro/src/tests/samples/default/src/Dependency.tsx
@@ -1,3 +1,0 @@
-export const Dependency = () => {
-  return <div className="Dependency">Hello, World!</div>;
-};
