--- pro/src/tests/samples/default/src/Other.tsx
@@ -1,6 +1,0 @@
-import React from "react";
-import "./App.css";
-
-export function Other() {
-  return <div className="Other">Hello, World!</div>;
-}
