--- pro/src/tests/samples/default/src/types.ts
@@ -1,3 +1,0 @@
-export interface Foo {
-  bar: string;
-}
